test: split test_telemetry_repository.py by test domain
The file was 872 lines, over the 800-line house-lint HSL102 threshold.
Split into three domain-grouped files with no test logic changes:

- test_telemetry_repository_registration.py — register/mark/upsert (272 lines)
- test_telemetry_repository_reconcile.py — reconciliation (375 lines)
- test_telemetry_repository_schema.py — persist batch, schema shape,
  query builders (241 lines)

The four module-level helpers (assert_listener_count, fetch_listener_field,
insert_committed_execution, insert_new_session) move to
_fixtures_telemetry.py and are re-exported from conftest.py, matching the
directory's existing convention rather than being duplicated per file.

All 44 definitions are byte-identical to the original; the split files
collect and pass the same 44 tests.

Closes #1758
