+++ tests/unit/core/CLAUDE.md
@@ -25,6 +25,10 @@ surface stable; only the definitions moved.
 - `make_watcher(hassette)`, `make_watcher_hassette(**kw)` — `ServiceWatcher` test setup
 - `make_blocking_io_hassette(**kw)` — minimal mock Hassette for watchdog and monkeypatch guard tests
 - `make_marker_executor(**kw)` — mock executor with `ExecutionMarker` on `current_execution`
+- `assert_listener_count(db, listener_id, expected, message)` — assert the number of `listeners` rows with that id
+- `fetch_listener_field(db, listener_id, field)` — read one column from a `listeners` row
+- `insert_committed_execution(db, session_id, **kw)` — insert and commit an `executions` row (1ms, now)
+- `insert_new_session(db)` — insert a second `running` session row and return its id, for the once=True/previous-session reconciliation tests
 
 ## Key conventions
